Ability to initialize a suit from a previuos run
This allows comparing from an older run,
this is useful when running on branches.

This method is quite slow though,
it first creates copies of all the baselines (including dupling the
image files) when the first run is created in a suit. Which would
increas the time-to-first-test with ~1min per 500tests.