Optimize compare, change IM to graphicsmagick
This does not convert the original images to fit them inside a canvas
size, because it takes ~100ms for imagemagic to do that for each image
(baseline and test). The new file was previously named .png, which now
requires a weird workaround, this should be fixable in some other way.

Graphicsmagic does not support the same compare options as imagemagic,
so this works very different.

Total processing (background job, depends on server load)

Before: ~3000ms
After: ~600ms